Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.3
botocore-stubs==1.37.29
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
dataclasses-json==0.6.7
datasets==3.6.0
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
griffe==1.7.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.60
langchain-openai==0.3.17
langsmith==0.3.42
Levenshtein==0.27.1
llama-cloud==0.1.19
llama-cloud-services==0.6.22
llama-index-core==0.12.37
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.40
llama-index-multi-modal-llms-openai==0.4.3
llama-index-readers-file==0.4.7
llama-index-tools-mcp==0.2.1
llama-index-tools-tavily-research==0.3.0
llama-parse==0.6.22
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy-boto3-bedrock-runtime==1.37.30
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.79.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
pyparsing==3.2.3
pypdf==5.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
sentry-sdk==2.25.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sse-starlette==2.3.5
standardwebhooks==1.0.0
starlette==0.46.2
striprtf==0.0.26
sympy==1.14.0
tavily-python==0.7.9
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
types-awscrt==0.26.1
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20241206
types-s3transfer==0.11.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0